Refactor randomNTimes and use ScopedTypeVariables
Now randomNTimes first generates a list of random generators, and only then maps over that list to get the final result of type a. This decouples the random generator logic from the function f.

In the process, we turn on my very first Haskell language extensions ever! ScopedTypeVariables allows for referring to `g` in `where` clauses.